requirements.txt
streamlit
scikit-learn
pandas
numpy
torch
torchvision
Pillow
xgboost
joblib
